Refactoring to allow AWS / provider config and account data to be passed in at runtime.
