+++ lib/billder-fetch-data.js
@@ -3,7 +3,8 @@
  * Use the 'aws-billing' package to get billing data for a specific account.
  */
 
-var config = require('../config');
+// The config filename has been passed in from the CLI
+//var config = require('./' + args[0]);
 
 module.exports = {
 
@@ -14,7 +15,7 @@ module.exports = {
    * @param {Boolean} withoutTaxes
    * @return {Object} AWS costs for provided linkedAccountId
    */
-  get_billing_data(month, linkedAccountId, withoutTaxes, callback) {
+  get_billing_data(config, month, linkedAccountId, withoutTaxes, callback) {
     var billing = require('aws-billing')(config.aws.accountId, config.aws.key, config.aws.secret, config.aws.bucket, config.aws.region, month, linkedAccountId, withoutTaxes);
     billing(function (err, costs) {
       return callback(costs);
