+++ lib/billder-make-invoice.js
@@ -3,12 +3,7 @@
  * Prepare an account object and pass it to the accounting provider that generates invoices.
  */
 
-const accounts = require('../accounts');
-const config = require('../config');
 const api = require('./billder-api-request');
-
-var provider_path = './provider/' + config.provider.name + '/';
-const make_provider_invoice = require(provider_path + 'billder-make-invoice');
 
 var syslog = require('modern-syslog');
 var request = require('request-promise');
@@ -18,7 +13,7 @@ var request = require('request-promise');
  * @param {String} month, 'YYYY-MM' string for target month for this invoice
  * @param {Object} account, object containing data about a customer who needs invoicing
  */
-function make_invoice(month, account) {
+function make_invoice(month, config_file, account) {
 
   /* 
    * Fetch the current exchange rate.
@@ -30,8 +25,12 @@ function make_invoice(month, account) {
       response_object = JSON.parse(response);
       // Stash the exchange rate in the account object so we don't get it overwritten by asynchronous calls later
       account['exchange_rate'] = response_object.rates[account.currency];
+      // Load provider from config file
+      var config = require('../' + config_file);
+      var provider_path = './provider/' + config.provider.name + '/';
+      var make_provider_invoice = require(provider_path + 'billder-make-invoice');
       // Send the account data to our provider-specific invoice building function
-      make_provider_invoice.make_invoice(month, account);
+      make_provider_invoice.make_invoice(month, config_file, account);
 
     })
     .catch(function (error) {
@@ -48,11 +47,15 @@ module.exports = {
    * Cycle through all accounts in ./accounts.js and trigger invoice processing.
    * @param {String} month, 'YYYY-MM' string for target month for these invoices
    */
-  make_invoices(month) {
+  make_invoices(month, config_file) {
+    // The config filename has been passed in from the CLI
+    var config = require('../' + config_file);
+    var accounts = require('../' + config.aws.accountsFile);
+    // Loop through accounts
     for (var account in accounts) {
       // Pass in customer name so we have it available for convenience
       accounts[account]['name'] = account;
-      make_invoice(month, accounts[account]);
+      make_invoice(month, config_file, accounts[account]);
     }
   }
