+++ lib/provider/freeagent/billder-auth.js
@@ -6,12 +6,6 @@
 const simpleOauthModule = require('simple-oauth2');
 var syslog = require('modern-syslog');
 
-// Set the configuration settings
-const config = require('../../../config');
-var credentials = {};
-credentials['client'] = config.provider.oauth.client;
-credentials['auth'] = config.provider.oauth.auth;
-
 module.exports = {
 
   /*
@@ -20,7 +14,14 @@ module.exports = {
    *
    * @param {String} access_token_file, path to file we write the token in
    */
-  authenticate(access_token_file='/tmp/fa-access-token.txt') {
+  authenticate(config_file, access_token_file='/tmp/fa-access-token.txt') {
+    // Load the config file
+    var config = require('../../../' + config_file);
+    // Set the configuration settings
+    var credentials = {};
+    credentials['client'] = config.provider.oauth.client;
+    credentials['auth'] = config.provider.oauth.auth;
+
     syslog.notice('Billder: refreshing OAuth access token');
     console.log('Billder: refreshing OAuth access token');
