+++ lib/provider/freeagent/billder-make-invoice.js
@@ -4,8 +4,6 @@
  */
 
 const billing_data = require('../../billder-fetch-data');
-const accounts = require('../../../accounts');
-const config = require('../../../config');
 const api = require('../../billder-api-request');
 
 var syslog = require('modern-syslog');
@@ -18,12 +16,17 @@ module.exports = {
    * @param {Object} account, object containing data about a customer who needs invoicing
    */
 
-  make_invoice(month, account) {
+  make_invoice(month, config_file, account) {
+
+    // Load in config
+    var config = require('../../../' + config_file);
 
     /*
      * Fetch billing data from AWS.
      */
-    billing_data.get_billing_data(month, account.AWS, true, function(costs) {
+    billing_data.get_billing_data(config, month, account.AWS, true, function(costs) {
+      var accounts = require('../../../' + config.aws.accountsFile);
+
       var invoice_items = [];
       var running_total = 0;
       var products = [];
@@ -115,7 +118,7 @@ module.exports = {
       max = Math.floor(20000);
       ms = Math.floor(Math.random() * (max - min)) + min; // The maximum is exclusive and the minimum is inclusive
       console.log('Billder: pausing generation of invoice for %s for %s milliseconds', account.name, ms);
-      setTimeout(create_invoice, ms, account, invoice);
+      setTimeout(create_invoice, ms, account, config, invoice);
 
     })
 
@@ -129,7 +132,7 @@ module.exports = {
  * @param {Object} account
  * @param {Object} invoice
  */
-function create_invoice(account, invoice) {
+function create_invoice(account, config, invoice) {
     options = api.build_request_options(config.provider.oauth.auth.tokenHost, '/v2/invoices', '/tmp/fa-access-token.txt', 'POST', invoice);
     request(options)
       .then(function (response) {
